requirements.txt
aiogram>=3.0.0
aiohttp>=3.9.0
trafilatura>=1.6.0
python-dotenv>=1.0.0
